+++ packages/client-connectivity/src/lib/config.ts
@@ -59,7 +59,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
     language: "typescript",
     openapiDocVersion: "0.0.0",
-    sdkVersion: "1.0.0-alpha.15",
-    genVersion: "2.385.2",
-    userAgent: "speakeasy-sdk/typescript 1.0.0-alpha.15 2.385.2 0.0.0 @formance/sdk-connectivity",
+    sdkVersion: "1.0.0-alpha.16",
+    genVersion: "2.387.0",
+    userAgent: "speakeasy-sdk/typescript 1.0.0-alpha.16 2.387.0 0.0.0 @formance/sdk-connectivity",
 } as const;