+++ packages/client-connectivity/RELEASES.md
@@ -159,3 +159,13 @@ Based on:
 - [typescript v1.0.0-alpha.15] packages/client-connectivity
 ### Releases
 - [NPM v1.0.0-alpha.15] https://www.npmjs.com/package/@formance/sdk-connectivity/v/1.0.0-alpha.15 - packages/client-connectivity
+
+## 2024-08-07 00:21:54
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.355.0 (2.387.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.16] packages/client-connectivity
+### Releases
+- [NPM v1.0.0-alpha.16] https://www.npmjs.com/package/@formance/sdk-connectivity/v/1.0.0-alpha.16 - packages/client-connectivity